backend/requirements.txt
fastapi
uvicorn
transformers
torch
sentencepiece
mlflow
pymongo
python-dotenv
